Move some files to registration_config_updater
