+++ lib/pokotarou/registration_config_updater/array_utils.rb
@@ -1,6 +1,6 @@
 class NothingDataError < StandardError; end
-class ArrayOperation
-  class << self
+module RegistrationConfigUpdater 
+  module ArrayUtils
     # return rotated val in passed array
     def get_rotated_val array, size, cnt
       raise NothingDataError.new("Nothing seed data") if array.nil?
